Defined the ItemList array-like interface and added a first set of unit tests
Spent some time testing and evaluating the available options to wrap the
ItemList class around a array-style interface. I tried subclassing,
using an explicit delegator object, but neither worked well enough
because the Array class isn't very 'inheritance' friendly (all methods
that return a new array-like interface use the Array class explicitly,
instead of the using the subclass). So I settled with the Forwardable
module, which forwards a select subset of methods to the wrapped array
instance.

To help with testing this, I've also (finally) added the first batch of
unit tests for the ItemList class.